Refactor `requestSongs`, `aggregateAndSave` and `songAggregator`
The reason for refactoring these three at once is so that now when the user requests a song, this flow of searching and saving to the database does not await `scrapeAmazonMusic` which caused the long response time.  Now, it aggregates and save data from spotify and deezer only, checks to see if there are duplicates, if not  saves it to the database; then sends the response whether it was saved or not then finally call `scrapeAmazonMusic` and try to update the song's price.

With this flow, the user will be able to receive the response without having to wait for the price from amazon. But it means that they won't always get the full data of the song the first time.